Move SimpleGameModel to common.
