+++ src/common/models/SimpleGameModel.ts
@@ -1,8 +1,8 @@
 
-import {Color} from '../common/Color';
-import {PlayerId, GameId, SpectatorId} from '../common/Types';
-import {Phase} from '../common/Phase';
-import {GameOptionsModel} from '../common/models/GameOptionsModel';
+import {Color} from '../Color';
+import {PlayerId, GameId, SpectatorId} from '../Types';
+import {Phase} from '../Phase';
+import {GameOptionsModel} from './GameOptionsModel';
 
 export interface SimpleGameModel {
     activePlayer: Color;
